requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.5.1
pymongo==4.8.0

# ── Cache ─────────────────────────────────────────────────────────────────────
redis==5.0.8
aioredis==2.0.1

# ── Exchange ──────────────────────────────────────────────────────────────────
ccxt==4.3.96

# ── HTTP ──────────────────────────────────────────────────────────────────────
httpx==0.27.2
aiohttp==3.10.5

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru==0.7.2

# ── Numerics / TA ─────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1

# ── Observability ─────────────────────────────────────────────────────────────
prometheus-client==0.20.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.1
uuid==1.30

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
pytest-timeout==2.3.1
